Bug 19846 - variable decl associated with the wrong debug info section
The decl of a member variable which DIE is in the main debug info
section is being associated with the alternate debug info section,
just because the specification DIE of the variable is in the alternate
debug info section.

This patch associates the decl of a member variable with the same
debug info section as its DIE, not with the debug info section of its
specification DIE.

Note that I am not adding a regression test for this because the
reproducer RPMs given in the bug for this issue are too big and slow
to analyze.

	* src/abg-dwarf-reader.cc (build_ir_node_from_die): Associate the
	decl of the variable to the same debug info file (alternate or
	not) as the DIE, not as its specification DIE.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>