Speedup diff node child insertion
Diff not chilrend nodes must remain sorted.  Before this patch, this
was achieved by sorting the vector of children each time a new member
was inserted.

But profiling showed that doing that for corpus_diff nodes was
relatively slow on corpora with lots function/variable changes. At
least enough to become a hot spot.

This patch addresses that by inserting the new child at the right
point in the vector of children, for the vector to remain sorted.

	* src/abg-comparison.cc (corpus_diff::append_child_node): Insert
	the new child at the right point in the vector of children, so
	that it remains sorted.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>