Avoid unnecessary computation of type name in suppression evaluation
When evaluating type suppression specifications, looking at the name
of the type can take a noticeable time, at least when the name of that
type is being queried for the first time.

Now that every single type involved in a change that might be reported
is involved in a type suppression evaluation (at least when operating
in the mode where the comparison tool is asked to filter out private
types), the evaluation of type suppression becomes a hot spot, at
least on relatively big diff graphs.

Luckily, private type suppression specifications don't involve the
name of the type, so we don't have to query the name of the type in
this case.

This patch makes it so that the type name is queried only when the
suppression specification requires it.  It makes us save a few
noticeable CPU cycles when evaluating suppression specifications that
don't involve looking at the type name.

	* src/abg-suppression.cc (type_suppression::suppresses_type):  If
	neither the type suppression "name" or "name_regex" properties
	where provided in the suppression specification, then do not try
	to look at the type name.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>